+++ fake-cucumber/javascript/test/TestCaseTest.ts
@@ -51,6 +51,7 @@ describe('TestCase', () => {
       await testCase.execute(
         (message: messages.Envelope) => emitted.push(message),
         0,
+        false,
         'test-case-started-id'
       )
 
@@ -59,6 +60,27 @@ describe('TestCase', () => {
 
       assert.strictEqual(testCaseStarted.testCaseId, testCase.id)
       assert.strictEqual(testCaseFinished.testCaseStartedId, testCaseStarted.id)
+    })
+
+    it('indicates an upcoming retry on TestCaseFinished when FAILED', async () => {
+      const emitted: messages.Envelope[] = []
+      const testSteps: ITestStep[] = [passedPickleTestStep]
+      const testCase = new TestCase(
+        'some-test-case-id',
+        testSteps,
+        'some-pickle-id',
+        new IncrementClock()
+      )
+      await testCase.execute(
+        (message: messages.Envelope) => emitted.push(message),
+        0,
+        true,
+        'test-case-started-id'
+      )
+
+      const testCaseFinished = emitted.find((m) => m.testCaseFinished).testCaseFinished
+
+      assert.strictEqual(testCaseFinished.willBeRetried, true)
     })
   })
 })