+++ messages/javascript/test/getWorstTestStepResultsTest.ts
@@ -8,17 +8,14 @@ describe('getWorstTestStepResult', () => {
       {
         status: TestStepResultStatus.PASSED,
         duration: { seconds: 0, nanos: 0 },
-        willBeRetried: false,
       },
       {
         status: TestStepResultStatus.FAILED,
         duration: { seconds: 0, nanos: 0 },
-        willBeRetried: false,
       },
       {
         status: TestStepResultStatus.PASSED,
         duration: { seconds: 0, nanos: 0 },
-        willBeRetried: false,
       },
     ])
     assert.strictEqual(result.status, TestStepResultStatus.FAILED)