+++ messages/ruby/lib/cucumber/messages.dtos.rb
@@ -1635,12 +1635,16 @@ module Cucumber
 
       attr_reader :timestamp
 
+      attr_reader :will_be_retried
+
       def initialize(
         test_case_started_id: '',
-        timestamp: Timestamp.new
+        timestamp: Timestamp.new,
+        will_be_retried: false
       )
         @test_case_started_id = test_case_started_id
         @timestamp = timestamp
+        @will_be_retried = will_be_retried
       end
     end
 
@@ -1786,18 +1790,14 @@ module Cucumber
 
       attr_reader :status
 
-      attr_reader :will_be_retried
-
       def initialize(
         duration: Duration.new,
         message: nil,
-        status: TestStepResultStatus::UNKNOWN,
-        will_be_retried: false
+        status: TestStepResultStatus::UNKNOWN
       )
         @duration = duration
         @message = message
         @status = status
-        @will_be_retried = will_be_retried
       end
     end